Use 1 ViewModel instead of having 2 for movies and moviedetails
