+++ app/src/main/java/com/example/top10movies/ui/viewmodels/MoviesViewModel.kt
@@ -9,6 +9,7 @@ import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
 import com.example.top10movies.Top10MoviesApp
+import com.example.top10movies.api.data.MovieDetails
 import com.example.top10movies.api.data.MoviesResult
 import com.example.top10movies.repo.MoviesRepo
 import com.example.top10movies.util.Constants.Companion.API_KEY
@@ -22,10 +23,12 @@ import java.io.IOException
 
 class MoviesViewModel(
     app: Application,
-    val moviesRepo: MoviesRepo
+    val moviesRepo: MoviesRepo,
+    //val movieDetailsRepo: MovieDetailsRepo
 ) : AndroidViewModel(app) {
 
-    val popularMovies: MutableLiveData<Resource<MoviesResult>> = MutableLiveData()
+    val movies: MutableLiveData<Resource<MoviesResult>> = MutableLiveData()
+    val movieDetails: MutableLiveData<Resource<MovieDetails>> = MutableLiveData()
 
     init {
         getPopularMovies(API_KEY)
@@ -46,18 +49,48 @@ class MoviesViewModel(
     }
 
     private suspend fun safeGetPopularMoviesCall(apiKey: String) {
-        popularMovies.postValue(Resource.Loading())
+        movies.postValue(Resource.Loading())
         try {
             if (hasInternetConnection()) {
                 val response = moviesRepo.getPopularMovies(apiKey)
-                popularMovies.postValue(handleMovieResult(response))
+                movies.postValue(handleMovieResult(response))
             } else {
-                popularMovies.postValue(Resource.Error(NO_INTERNET))
+                movies.postValue(Resource.Error(NO_INTERNET))
             }
         } catch (t: Throwable) {
             when(t) {
-                is IOException -> popularMovies.postValue(Resource.Error(NETWORK_FAILURE))
-                else -> popularMovies.postValue(Resource.Error(JSON_PARSING_ERROR))
+                is IOException -> movies.postValue(Resource.Error(NETWORK_FAILURE))
+                else -> movies.postValue(Resource.Error(JSON_PARSING_ERROR))
+            }
+        }
+    }
+
+    fun getMovieDetailsById(apiKey: String, movieId: String) = viewModelScope.launch {
+        safeMovieDetailsCall(apiKey, movieId)
+    }
+
+    private fun handleMovieDetailsResult(response: Response<MovieDetails>) : Resource<MovieDetails> {
+        if (response.isSuccessful) {
+            response.body()?.let { resultResponse ->
+                return Resource.Success(resultResponse)
+            }
+        }
+        return Resource.Error(response.message())
+    }
+
+    private suspend fun safeMovieDetailsCall(apiKey: String, movieId: String) {
+        movieDetails.postValue(Resource.Loading())
+        try {
+            if (hasInternetConnection()) {
+                val response = moviesRepo.getMovieDetailsById(apiKey, movieId)
+                movieDetails.postValue(handleMovieDetailsResult(response))
+            } else {
+                movieDetails.postValue(Resource.Error(NO_INTERNET))
+            }
+        } catch (t: Throwable) {
+            when(t) {
+                is IOException -> movieDetails.postValue(Resource.Error(NETWORK_FAILURE))
+                else -> movieDetails.postValue(Resource.Error(JSON_PARSING_ERROR))
             }
         }
     }
